initial commit for project changes
ATM Interface
- There are five different classes in the project, namely, account holder, account, bank transaction, bank
and particular ATM of the bank. This project is a console-based application. If you start running the 
program, you will be prompted with user id and user pin. If you entered it successfully, then you
unlock all the functionalities which typically exist in an ATM. The operations you can perform in this
project are show transactions history, withdraw, deposit, transfer and quit

 Simple Banking Application
- In simple banking application learner will be coding the simple bank operations like check balance,
deposit, withdraw, exit, etc. In this project learner will be learning how to take input from the user
using Scanner class, basics about String, how to print in java, variables, if/else statements, methods,
Loops etc..